requirements.txt
streamlit>=1.24.0
numpy>=1.24.0
opencv-python>=4.8.0
pillow>=9.5.0
scipy>=1.10.0
matplotlib>=3.7.0
scikit-image>=0.21.0
pyyaml>=6.0.0
streamlit-image-select>=0.6.0
streamlit-image-coordinates>=0.1.3